requirements.txt
requests
pandas
python-dotenv
openpyxl
streamlit
GitPython